Make unit tests on processor inherit from a parent test class
