+++ QueryBuilder_Cassandra/src/test/java/net/sf/esfinge/querybuilder/cassandra/unit/querybuilding/resultsprocessing/SecondaryQueryProcessorTest.java
@@ -1,0 +1,6 @@
+package net.sf.esfinge.querybuilder.cassandra.unit.querybuilding.resultsprocessing;
+
+public class SecondaryQueryProcessorTest extends BasicProcessorTest {
+
+
+}
