+++ QueryBuilder_Cassandra/src/test/java/net/sf/esfinge/querybuilder/cassandra/unit/querybuilding/resultsprocessing/SpecialComparisonProcessorTest.java
@@ -5,7 +5,6 @@ import net.sf.esfinge.querybuilder.cassandra.querybuilding.resultsprocessing.Spe
 import net.sf.esfinge.querybuilder.cassandra.querybuilding.resultsprocessing.specialcomparison.SpecialComparisonClause;
 import net.sf.esfinge.querybuilder.cassandra.querybuilding.resultsprocessing.specialcomparison.SpecialComparisonType;
 import net.sf.esfinge.querybuilder.cassandra.unit.reflection.TestClass;
-import org.junit.Before;
 import org.junit.Test;
 
 import java.util.ArrayList;
@@ -13,26 +12,7 @@ import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class SpecialComparisonProcessorTest {
-
-    private List<TestClass> objectList;
-
-    @Before
-    public void init() {
-        objectList = new ArrayList<>();
-
-        TestClass obj1 = new TestClass(1, "Pedro", "Silva");
-        TestClass obj2 = new TestClass(2, "Marcos", "Ferreira");
-        TestClass obj3 = new TestClass(3, "Antonio", "Marques");
-        TestClass obj4 = new TestClass(4, "Marcos", "Silva");
-        TestClass obj5 = new TestClass(5, "Silvia", "Bressan");
-
-        objectList.add(obj1);
-        objectList.add(obj2);
-        objectList.add(obj3);
-        objectList.add(obj4);
-        objectList.add(obj5);
-    }
+public class SpecialComparisonProcessorTest extends BasicProcessorTest {
 
     @Test
     public void filterListNotEqualWithOneParameterNotEqualTest() {
